% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automagic.R
\name{automagic}
\alias{automagic}
\title{Automagically install all required R packages}
\usage{
automagic(directory = getwd())
}
\arguments{
\item{directory}{folder to search for R and Rmd files}
}
\description{
Searches a given directory for all R and R Markdown files, parses them for
required packages and attempts to install them from CRAN or GitHub. More importantly,
if a `deps.yaml` file was made using \code{\link{make_deps_file}}, automagic
will use this rather than try to install based on a best guess.
}
\seealso{
\code{\link{install_package_guess}}, \code{\link{parse_packages}}
}
