% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dependent_packages.R
\name{get_dependent_packages}
\alias{get_dependent_packages}
\title{get packages required to run R code}
\usage{
get_dependent_packages(directory = getwd())
}
\arguments{
\item{directory}{folder to search for R and Rmd files}
}
\value{
a vector of package names
}
\description{
get packages required to run R code
}
\details{
parses all R and Rmd files in a directory and uses \code{\link{parse_packages}}
    to find all R packages required for the code to run
}

