% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_details.R
\name{get_package_details}
\alias{get_package_details}
\title{get package details}
\usage{
get_package_details(pkg_name)
}
\arguments{
\item{pkg_name}{package name}
}
\value{
A list of package characteristics.
  "Package", "Repository", and "Version" for CRAN packages.
  "Package", "GithubUsername", "GithubRepo", "GithubRef", and "GithubSHA1" for Github packages.
}
\description{
Uses \code{packageDescription} to get details about given package from R library on local machine.
Currently only supports CRAN and GitHub packages
}
