% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysreqs.R
\name{get_sysreqs}
\alias{get_sysreqs}
\title{Get system dependencies for R packages}
\usage{
get_sysreqs(pkgs)
}
\arguments{
\item{pkgs}{a character vector of package names}
}
\description{
Calls the r-hub "sysreqs" API. Platform is currently hardcoded for Ubuntu
16.04 (\code{linux-x86_64-debian-gcc})
}
