% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps_file.R
\name{install_deps_file}
\alias{install_deps_file}
\title{Install R packages from a package dependencies (\code{deps.yaml}) file}
\usage{
install_deps_file(directory = getwd())
}
\arguments{
\item{directory}{directory containing \code{deps.yaml} file}
}
\description{
Installs packages from GitHub and CRAN based on Sha1 key and version number
respectively, as defined in a \code{deps.yaml} file created by
\code{\link{make_deps_file}}
}
\seealso{
\code{\link{make_deps_file}}, \code{\link{automagic}}
}

