% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_package.R
\name{install_package}
\alias{install_package}
\title{Install package from CRAN or GitHub}
\usage{
install_package(pkg, force_install = FALSE)
}
\arguments{
\item{pkg}{name of package to install}

\item{force_install}{install even if package is in library}
}
\description{
Install package from CRAN or GitHub
}
\details{
If a package is not availble in the R library, attempt to install it
  from CRAN.  If not available on CRAN, attempt to install the package from
  GitHub based on the "best guess".

  This function does not check package versions.  Specify
  \code{force_install=TRUE} to force installation of the package, updating it
  to the latest available version. Note that this function attempts to
  install its packages based on a best guess and is meant for use in
  automatically setting up an R programming environment. Do not use for
  installing packages if you have the option to install with other tools like
  \code{\link[utils]{install.packages}} or one of the various GitHub
  installers.
}
\examples{
\dontrun{
install_package('utils')
}

}
\seealso{
\code{\link{parse_packages}}, \code{\link{automagic}}
}

