% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps_file.R
\name{make_deps_file}
\alias{make_deps_file}
\title{Make a package dependencies (\code{deps.yaml}) file}
\usage{
make_deps_file(directory = getwd())
}
\arguments{
\item{directory}{directory containing R code to parse}
}
\description{
This function parses R code for required packages using
 \code{\link{parse_packages}} and then queries the R package library to
 determine the exact source and version of each package to install.
 Currently, only CRAN and GitHub packages are supported.
 Install packages from the `deps.yaml` file using
 \code{\link{automagic}{install_deps_file}}
}
\seealso{
\code{\link{automagic}}
}
