% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_dockerize.R
\name{shiny_dockerize}
\alias{shiny_dockerize}
\title{dockerize a shiny application}
\usage{
shiny_dockerize(directory = getwd(), re.automagic = FALSE,
  app.name = basename(directory))
}
\arguments{
\item{directory}{path to directory containing the shiny app (defaults to working directory)}

\item{re.automagic}{logical; force automagic to recreate dependencies file}

\item{app.name}{defaults to the basename of \code{directory}}
}
\description{
dockerize a shiny application
}

