% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_description_file.R
\name{update_description_file}
\alias{update_description_file}
\title{add missing packages to package DESCRIPTION file}
\usage{
update_description_file(directory = getwd())
}
\arguments{
\item{directory}{folder to search for R and Rmd files.
The DESCRIPTION file is expected to be in this directory.}
}
\description{
add missing packages to package DESCRIPTION file
}
\details{
uses \code{get_dependent_packages()} to find dependencies and
adds them to the description file with the desc package
}
