% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automagic.R
\name{automagic}
\alias{automagic}
\title{Automagically install all required R packages}
\usage{
automagic(directory = getwd(), ...)
}
\arguments{
\item{directory}{folder to search for R and Rmd files}

\item{...}{other arguments to \code{\link[automagic]{install_package}},
namely \code{force_install}}
}
\value{
invisibly returns a list of packages that were found
}
\description{
Searches a given directory for all R and R Markdown files, parses them for
required packages and attempts to install them from CRAN or GitHub.
}
\details{
Used for autosetup of an R programming environment. For example,
  when running an R script on a remote server, or distributing an R Shiny app to
  another user, or creating a Docker container to run R code.  Just add
  \code{automagic::automagic()} to the top of the R file.
}
\examples{
# not run
automagic()
}
\seealso{
\code{\link{install_package}}, \code{\link{parse_packages}}
}

