% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps_file.R
\name{install_deps_file}
\alias{install_deps_file}
\title{Install R packages from a .dependencies file}
\usage{
install_deps_file(directory = getwd())
}
\arguments{
\item{directory}{directory containing .dependencies file}
}
\description{
Install R packages from a .dependencies file
}
\details{
Installs packages from GitHub and CRAN based on Sha1 key and version number
respectively, as defined in a .dependencies file created by
\code{\link[automagic]{make_deps_file}}
#' @seealso \code{\link{automagic}}
}

