% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps_file.R
\name{make_deps_file}
\alias{make_deps_file}
\title{Make a .dependencies file}
\usage{
make_deps_file(directory = getwd())
}
\arguments{
\item{directory}{directory containing R code to parse}
}
\description{
Make a .dependencies file
}
\details{
This function parses R code and then queries the R package library to
determine the exact source and version of each package to install. Currently,
only CRAN and GitHub packages are supported using a version number and Sha1
key, respectively. Install packages from the .dependencies file using
\code{\link[automagic]{install_deps_file}}
}
\seealso{
\code{\link{automagic}}
}

